Fix bug that prevented shutdown on Ctrl+C
In #62 the signal handler uses the default options, which will restart
interrupted system calls. This disables that behavior so the server will
shut down properly.

Additionally, the global socket_fd variable was shadowed by a local copy
in main() so it was never being closed. Removed the redundant shutdown,
refactored out the global.