Implement alpha-quality support for protocol buffers.
This adapts the chat example to use protocol buffers for communication.
There are still pieces that have not been migrated (heartbeats) but this
represents a reasonable starting point that we can work from.

Long term, we need to remove the remnants of the old messaging format to
simplify the codebase.